+++ src/pywindow/_internal/__init__.py
@@ -1,0 +1,1 @@
+"""pywindow module."""
