+++ src/pywindow/_internal/io_tools.py
