+++ src/pywindow/_internal/tables.py
