+++ src/pywindow/_internal/trajectory.py
