+++ src/pywindow/_internal/utilities.py
